pull out addon specific code into their own files
